update the definition of the coordinates.
Remove LightTools coordinates. A simple x-y coordinate will be
used for focal information and the spherical coordinate will be
used for beam directions. The global coordinate is defined that
z-axis to be the zenith of GB and x-axis to be the direction of
the focal plane from the secondary mirror.